Added in new map command
